requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.4.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncio==3.4.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
category-encoders==2.6.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
cmdstanpy==1.2.5
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
contractions==0.1.73
cycler==0.12.1
cymem==2.0.10
Cython==3.0.11
dacite==1.8.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
es_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.8.0/es_core_news_sm-3.8.0-py3-none-any.whl#sha256=e451a83d6df79b87e9eed0cb553f03e99e36a3bab18a7b79f0dcfd1fdf875e12
et_xmlfile==2.0.0
executing==2.1.0
Faker==33.3.1
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.1.0
Flask-JWT-Extended==4.7.1
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
gotrue==2.11.1
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h2==4.1.0
h5py==3.12.1
holidays==0.63
hpack==4.0.0
htmlmin==0.1.12
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.5
hyperframe==6.0.1
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.12.4
imblearn==0.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.27.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.2
jupyter_nbextensions_configurator==0.6.4
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.7.0
keras-tuner==1.4.7
kiwisolver==1.4.7
kt-legacy==1.0.5
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multimethod==1.12
murmurhash==1.0.11
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.14.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.59.5
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
phik==0.12.4
pillow==10.4.0
platformdirs==4.3.2
plotly==5.24.1
plotly-express==0.4.1
pmdarima==2.0.4
postgrest==0.19.1
preshed==3.0.9
prometheus_client==0.21.0
prompt_toolkit==3.0.47
propcache==0.2.1
prophet==1.1.6
protobuf==5.28.3
psutil==6.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyahocorasick==2.1.0
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.10.1
pyod==2.0.2
pyparsing==3.1.4
pyproj==3.7.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
PyWavelets==1.7.0
pywin32==306
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
realtime==2.1.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
selenium==4.21.0
Send2Trash==1.8.3
sentinelhub==3.11.1
setuptools==75.3.0
shap==0.46.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.4
storage3==0.11.0
streamlit==1.40.1
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.2
sympy==1.13.1
tangled-up-in-unicode==0.2.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
textsearch==0.0.24
tf_keras==2.18.0
thinc==8.3.2
threadpoolctl==3.5.0
tifffile==2024.12.12
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.1.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.47.0
trio==0.26.2
trio-websocket==0.11.1
typeguard==4.4.1
typer==0.15.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
utm==0.7.0
visions==0.7.6
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wordcloud==1.9.3
wrapt==1.17.0
wsproto==1.2.0
xgboost==2.1.2
yarl==1.18.3
ydata-profiling==4.12.0
yellowbrick==1.5
